+++ header.php
@@ -117,7 +117,9 @@
 		<div class="group">
 			<div class="header-back" onclick="history.back()"><i class="fas fa-chevron-left"></i></div>
 			<a href="<?php echo esc_url(home_url('/')); ?>"><img src="<?php echo LOOPIS_THEME_HQ_URI; ?>/assets/img/LOOPIS_logo.png" alt="LOOPIS-logo" id="header-img"></a>
-				<div class="header-faq" onclick="location.href='<?php echo esc_url( home_url('/faq/') ); ?>'">💡</div>
+				<?php if (is_user_logged_in()) : ?>
+				<div class="header-faq" onclick="location.href='<?php echo esc_url( home_url('/12845/') ); ?>'">📍</div>
+				<?php endif; ?>
 			</div>
 		</header>