+++ page-shop.php
@@ -1,6 +1,8 @@
 <?php
 /**
  * Dynamic content for pages using url /shop/?option=
+ * 
+ * Content is fetched from "LOOPIS Theme"
  */
 ?>
 
@@ -8,11 +10,11 @@
 
 <div class="page-padding center">
 <?php if ( is_user_logged_in() ) : 
-    // Dynamic page loader
-    $page_dir = get_template_directory() . '/pages/shop/';
+    // Shared dynamic page loader from LOOPIS Theme.
+    $page_dir = LOOPIS_THEME_DIR . '/pages/shop/';
 
     // Get the 'option' parameter from URL
-    $page_option = isset($_GET['option']) ? sanitize_file_name($_GET['option']) : 'coins';
+    $page_option = isset($_GET['option']) ? sanitize_file_name($_GET['option']) : 'start';
 
     $php_file = $page_dir . $page_option . '.php';
 
@@ -20,14 +22,14 @@
         include $php_file;
     } else {
         echo '<h1>🛒 Shoppen</h1><hr>';
-        include LOOPIS_THEME_HQ_DIR . '/templates/access/loopis-404.php';
+        include LOOPIS_THEME_DIR . '/includes/output/access/loopis-404.php';
     }
     ?>
     <div class="clear"></div>
 
 <?php else :
 // Not logged in message
-include LOOPIS_THEME_HQ_DIR . '/templates/access/role-message.php';
+include LOOPIS_THEME_HQ_DIR . '/includes/output/access/role-message.php';
 endif; ?>
 
 </div><!--page-padding center-->