+++ page-user.php
@@ -1,6 +1,8 @@
 <?php
 /**
  * Dynamic content for pages using url /user/?option=
+ * 
+ * Content is fetched from "LOOPIS Theme"
  */
 ?>
 
@@ -9,8 +11,8 @@
 <div class="page-padding center">
     <h1>👤 Min profil</h1>
 <?php if ( is_user_logged_in() ) : 
-    // Dynamic page loader
-    $page_dir = get_template_directory() . '/pages/user/';
+    // Shared dynamic page loader from LOOPIS Theme.
+    $page_dir = LOOPIS_THEME_DIR . '/pages/user/';
 
     // Get the 'option' parameter from URL
     $page_option = isset($_GET['option']) ? sanitize_file_name($_GET['option']) : 'tabs';
@@ -21,14 +23,16 @@
         include $php_file;
     } else {
         echo '<h1>👤 Min profil</h1><hr>';
-        include LOOPIS_THEME_HQ_DIR . '/templates/access/loopis-404.php';
+        include LOOPIS_THEME_DIR . '/includes/output/access/loopis-404.php';
     }
     ?>
     <div class="clear"></div>
 
 <?php else :
 // Not logged in message
-include LOOPIS_THEME_HQ_DIR . '/templates/access/role-message.php';
+echo '<hr>';
+include LOOPIS_THEME_DIR . '/includes/output/access/only-user.php';
+include LOOPIS_THEME_HQ_DIR . '/includes/output/access/role-options.php';
 endif; ?>
 
 </div><!--page-padding center-->